+++ app/src/main/java/geekbrains/ru/banananotes/model/provider/RemoteDataProvider.kt
@@ -1,6 +1,9 @@
-package geekbrains.ru.banananotes.model
+package geekbrains.ru.banananotes.model.provider
 
 import androidx.lifecycle.LiveData
+import geekbrains.ru.banananotes.model.Note
+import geekbrains.ru.banananotes.model.NoteResult
+import geekbrains.ru.banananotes.model.User
 
 interface RemoteDataProvider {
     fun subscribeToAllNotes(): LiveData<NoteResult>
