+++ app/src/main/java/geekbrains/ru/banananotes/ui/viewstate/BaseViewState.kt
@@ -1,4 +1,4 @@
-package geekbrains.ru.banananotes.ui
+package geekbrains.ru.banananotes.ui.viewstate
 
 open class BaseViewState<T>(val data: T, val error: Throwable?) {
 }
