+++ app/src/main/java/geekbrains/ru/banananotes/ui/viewstate/SplashViewState.kt
@@ -1,4 +1,6 @@
-package geekbrains.ru.banananotes.ui
+package geekbrains.ru.banananotes.ui.viewstate
+
+import geekbrains.ru.banananotes.ui.viewstate.BaseViewState
 
 class SplashViewState(isAuth: Boolean? = null, error: Throwable? = null) :
     BaseViewState<Boolean?>(isAuth, error)
