+++ app/src/main/java/geekbrains/ru/banananotes/viewmodel/BaseViewModel.kt
@@ -3,7 +3,7 @@ package geekbrains.ru.banananotes.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import geekbrains.ru.banananotes.ui.BaseViewState
+import geekbrains.ru.banananotes.ui.viewstate.BaseViewState
 
 open class BaseViewModel<T, VS : BaseViewState<T>>() : ViewModel() {
     open val viewStateLiveData = MutableLiveData<VS>()
