+++ app/src/main/java/geekbrains/ru/banananotes/viewmodel/SplashViewModel.kt
@@ -1,8 +1,8 @@
 package geekbrains.ru.banananotes.viewmodel
 
 import geekbrains.ru.banananotes.model.NoAuthException
-import geekbrains.ru.banananotes.model.Repository
-import geekbrains.ru.banananotes.ui.SplashViewState
+import geekbrains.ru.banananotes.model.repository.Repository
+import geekbrains.ru.banananotes.ui.viewstate.SplashViewState
 
 class SplashViewModel(private val repository: Repository = Repository) :
     BaseViewModel<Boolean?, SplashViewState>() {
